+++ bcts/lrm/forest/ddl/bcts_lrm_forest_timber_inventory_ready_to_develop_ddl.sql
@@ -68,6 +68,6 @@ CREATE TABLE IF NOT EXISTS bcts_staging.timber_inventory_ready_to_develop_hist
     ogs_reactivated_re_engineered_status text COLLATE pg_catalog."default",
     ogs_reactivated_re_engineered date,
     spatial_flag character varying(3) COLLATE pg_catalog."default",
-    cutb_seq_nbr numeric(15,0),
+    cutb_seq_nbr bigint,
     report_end_date date
 )
